Fixed a visualization issue.
The visualization module used to plot image points on the canvas.
However, in the image ref frame y points down, while in the cavas ref
frame y points up. As a result the rendered ref frames show up as
left-handed. This could be easily fixed by manually converting all the
points to the canvas ref frame.